Define requirements web interface with data classes
